Extract node comparison method to trait
